requirements.txt
aioamqp==0.13.0
aioredis==1.3.0
alabaster==0.7.12
amqp==2.5.1
asgi-redis==1.4.3
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-rabbitmq==1.1.5
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-betterforms==1.2
django-celery-beat==1.5.0
django-import-export==2.0.1
django-rosetta==0.9.3
django-timezone-field==3.1
djangorestframework==3.10.3
doc8==0.8.0
docutils==0.16
et-xmlfile==1.0.1
flower==0.9.3
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.0
kombu==4.6.5
lazy-object-proxy==1.4.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.4
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.1
pamqp==2.3.0
pbr==5.4.4
pep8==1.7.1
Pillow==6.1.0
polib==1.1.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.3.1
PyMySQL==0.10.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.3
redis==2.10.6
reportlab==3.5.31
requests==2.22.0
restructuredtext-lint==1.3.0
rstcheck==3.3.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
stevedore==1.31.0
tablib==0.14.0
tornado==5.1.1
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0
